fix/pipeline-crudhandlertests

+++ api/DTOs/Mqtt/UpdateTemperatureDto.cs
@@ -4,10 +4,10 @@ public class UpdateTemperatureDto
 {
     public string Name { get; }
     public string Office_name { get; }
-    public float Desired_temp { get; }
+    public double Desired_temp { get; }
     public bool Window_toggle { get; }
 
-    public UpdateTemperatureDto(string name, string office_name, float desired_temp, bool window_toggle)
+    public UpdateTemperatureDto(string name, string office_name, double desired_temp, bool window_toggle)
     {
         Name = name;
         Office_name = office_name;
